Fix stacked history changes
When flicking through the history quickly, this waits for current
animations to complete before doing a new jump. Previously, jumps could
be started while animations for other jumps were in progress,
potentially causing multiple layouts being visible at once.

This could probably be generalised for all animations but it's difficult
to cause the buggy behaviour in all other places, so I've just done it
for the history for now. A nice thing about the history is that the
animations are all started from the original UI, so we don't have to
worry about animation queues in different UIs, which would complicate
this a lot.

I moved the animation queue stuff into a class mostly for debugging, so
I could give queues a name and log the steps. It could be reverted back
to an array if preferred but I don't think there are any downsides to
this move.