Browser history support (#258)
* First working history

* Move history initialization to setupHistory()

* Exclusively use browser history.

* Clear frames when moving to and from nodes through history.

* Call `parseHash` on the current layout

Might want to rename this to `jumpToHash` or something

* Debug logs

* Use animation queue when jumping up

* Rename parseHash to jumpToHash

* Run animation queue at the end of jumpToCollapsedNodeHash

* Fix jumping to top-level collapsed nodes (not nested in a cluster node)

* Move setupHistory() call to complete() where the other history stuff already is

* Fix stacked history changes

When flicking through the history quickly, this waits for current
animations to complete before doing a new jump. Previously, jumps could
be started while animations for other jumps were in progress,
potentially causing multiple layouts being visible at once.

This could probably be generalised for all animations but it's difficult
to cause the buggy behaviour in all other places, so I've just done it
for the history for now. A nice thing about the history is that the
animations are all started from the original UI, so we don't have to
worry about animation queues in different UIs, which would complicate
this a lot.

I moved the animation queue stuff into a class mostly for debugging, so
I could give queues a name and log the steps. It could be reverted back
to an array if preferred but I don't think there are any downsides to
this move.

* lint :P

* Clear historyAnimationQueue after queued jump

* Fix lingering animation queue after a queued jump

* Remove console.logs

* Unobfuscate this name

* Remove the default animation queue naming

* Fix history animations hanging when a jump was made from # to no hash

* some animation logging again

* Use a global queue for all animations

All `selectNode`, `jumpToNode`, and `traverseUp` calls have to be
wrapped in a `queueAnimation()` call, else we run into the same sync
issues as we did with the history animations. i.e. the selectedDataNode
and such being updated before a previous animation ended

* Support multiple AnimationQueue onComplete handlers

The `onComplete` assignment when selecting aggregate nodes
overwrote the one in `queueAnimation`, causing the queue to hang.

* Use queueing for initial animation, fix moving back to initial state

* Fix hanging when executing a breadcrumb animation to a UI that no longer exists

* Separate animation sequencing and on('complete')

These have some different use cases. Sequencing only has a single
handler, doing things on completion does not have to be a single
handler.

* Remove animation debug logs.