Fix native stack trace test
In v8 7.0, Array.sort() was moved from a JS-based implementation to
Torque:
https://github.com/v8/v8/commit/fa11e2ac0380948a2f57963e6651402c3ca07d04

This means Array.sort() isn't in v8's `array.js` file anymore.
Apparently the goal is to eventually move all or most JS implementations
of builtins like Array to to Torque, so 'native' is a shrinking
category. We may need to do something to detect those Torque things?